Separated types defs file into different type defs
