+++ client/src/types/PlanLineTypes.tsx
@@ -1,0 +1,12 @@
+import { PlanMarkerProps } from "./PlanMarkerTypes"
+
+interface PlanLineProps {
+    place1?: PlanMarkerProps
+    place2?: PlanMarkerProps
+    travelMode: string
+}
+
+interface PlanLineState {
+    place1: PlanMarkerProps
+    place2: PlanMarkerProps
+}
