+++ client/src/types/PlanMarkerTypes.tsx
@@ -1,0 +1,12 @@
+interface PlanMarkerProps {
+    title?: string
+    latitude?: number
+    longitude?: number
+}
+
+interface PlanMarkerState {
+    title: string,
+    location : google.maps.LatLng
+}
+
+export type { PlanMarkerProps, PlanMarkerState }
