+++ client/src/types/SearchResTypes.tsx
@@ -1,0 +1,13 @@
+interface SearchResProps {
+    title: string
+    desc: string
+    isOpen: boolean
+    addr: string
+    img?: string
+}
+
+interface SearchResState {
+
+}
+
+export type { SearchResProps, SearchResState }
